+++ packages/protocol/src/index.ts
@@ -25,3 +25,5 @@ export * from './ratelimit.js';
 export * from './replay.js';
 export * from './responses.js';
 export * from './sourcemap.js';
+export * from './ssrf.js';
+export * from './trackers.js';